requirements.txt
aiohttp==3.9.1
python-dotenv==1.0.0
slack-sdk==3.26.1
openai==1.6.1